+++ parsedmarc/__init__.py
@@ -1906,7 +1906,7 @@ def get_dmarc_reports_from_mailbox(
                 i + 1, message_limit, msg_uid
             )
         )
-        if isinstance(mailbox, MSGraphConnection):
+        if isinstance(connection, MSGraphConnection):
             if test:
                 msg_content = connection.fetch_message(msg_uid, mark_read=False)
             else:
@@ -2053,14 +2053,17 @@ def get_dmarc_reports_from_mailbox(
         ]
     )
 
-    if current_time:
-        total_messages = len(
-            connection.fetch_messages(reports_folder, since=current_time)
-        )
+    if not test and not batch_size:
+        if current_time:
+            total_messages = len(
+                connection.fetch_messages(reports_folder, since=current_time)
+            )
+        else:
+            total_messages = len(connection.fetch_messages(reports_folder))
     else:
-        total_messages = len(connection.fetch_messages(reports_folder))
+        total_messages = 0
 
-    if not test and not batch_size and total_messages > 0:
+    if total_messages > 0:
         # Process emails that came in during the last run
         results = get_dmarc_reports_from_mailbox(
             connection=connection,